Refactor document tools into modular structure with separate files
