+++ src/tools/document/children/index.ts
@@ -1,0 +1,2 @@
+export { docxV1DocumentBlockChildrenCreateSimple } from './create';
+export { docxBlockBatchDelete } from './delete';
