+++ src/tools/document/import/index.ts
@@ -1,0 +1,1 @@
+export { docxMarkdownImport } from './markdown';
