+++ src/tools/document/index.ts
@@ -1,11 +1,11 @@
 import z from 'zod';
-import { Client } from '@larksuiteoapi/node-sdk';
-import * as lark from '@larksuiteoapi/node-sdk';
 import { zodToJsonSchema } from 'zod-to-json-schema';
 
-import { FEISHU_CONSTANTS } from '../../config/feishu-constants';
-import { addMermaidBlockMarkers } from '../../utils/markdown-processor';
 export { blockTreeTool, docxBlockPatch } from './block';
+export { docxV1DocumentBlockChildrenCreateSimple, docxBlockBatchDelete } from './children';
+export { docxV1DocumentTableCreate } from './table';
+export { docxImageOrVideoOrFileCreate } from './media';
+export { docxMarkdownImport } from './import';
 
 
 
@@ -5376,647 +5376,11 @@ export const docxV1BlockTypeSchemaGet = {
   },
 };
 
-export const docxV1DocumentBlockChildrenCreateSimple = {
-  project: 'docx',
-  name: 'docx_block_create',
-  sdkName: 'docx.v1.documentBlockChildren.create',
-  path: '/open-apis/docx/v1/documents/:document_id/blocks/:block_id/children',
-  httpMethod: 'POST',
-  description:
-    '[Feishu/Lark]-云文档-文档-块-创建块-指定需要操作的块，为其创建一批子块，并插入到指定位置。如果操作成功，接口将返回新创建子块的富文本内容',
-  accessTokens: ['tenant', 'user'],
-  schema: {
-    data: z.object({
-      children: z.array(z.record(z.unknown())).describe('添加的块列表').optional(),
-      index: z.number().describe('当前 block 在 children 中的插入位置，起始值为 0，最大值为原 children 长度').optional(),
-    }),
-    path: z.object({
-      document_id: z.string().describe('文档唯一标识'),
-      block_id: z.string().describe('Block 的唯一标识，确定要在分栏、表格、高亮中创建子块，请传入分栏、表格、高亮的 BlockID，否则传入document_id'),
-    }),
-  },
-  customHandler: async (client: Client, params: any, options: any) => {
-    try {
-      if (params.data.children[0].block_type === 40 && params.data.children[0].mermaid_drawing) {
-        params = get补全后的文本绘图块参数(params);
-      } else if (params.data.children[0].block_type === 40 && params.data.children[0].glossary) {
-        params = get补全后的名词解释块参数(params);
-      } else if (params.data.children[0].block_type === 40 && params.data.children[0].timeline) {
-        params = get补全后的时间轴块参数(params);
-      } else if (params.data.children[0].block_type === 40 && params.data.children[0].catalog_navigation) {
-        params = get补全后的目录导航块参数(params);
-      } else if (params.data.children[0].block_type === 40 && params.data.children[0].information_collection) {
-        params = get补全后的信息收集块参数(params);
-      } else if (params.data.children[0].block_type === 40 && params.data.children[0].countdown) {
-        params = get补全后的倒计时块参数(params);
-      } else if (params.data.children[0].block_type === 23) {
-        params = get补全后的附件块参数(params);
-      }
-      const response = await client.docx.v1.documentBlockChildren.create(params, lark.withUserAccessToken(options.userAccessToken));
-      switch (params.data.children[0].block_type) {
-        case 27:
-          const imageBlockId = response.data?.children?.[0]?.block_id;
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: `图片块创建成功，接下来使用Image BlockID ${imageBlockId} 作为 parent_node 上传图片，上传完成后修改图片块`,
-              },
-            ],
-          };
-        case 23:
-          const fileBlockId = response.data?.children?.[0]?.children?.[0];
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: `文件块创建成功，接下来使用File BlockID ${fileBlockId} 作为 parent_node 上传文件，上传完成后修改文件块`,
-              },
-            ],
-          };
-        default:
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: JSON.stringify(response),
-              },
-            ],
-          };
-      }
-    } catch (error) {
-      console.error('docxV1DocumentBlockChildrenCreateSimple 工具执行失败:', error);
-      return {
-        isError: true,
-        content: [
-          {
-            type: 'text',
-            text: `docxV1DocumentBlockChildrenCreateSimple 工具执行失败: ${error instanceof Error ? error.message : '未知错误'}`,
-          },
-        ],
-      };
-    }
-  },
-};
 
-export const docxV1DocumentTableCreate = {
-  project: 'docx',
-  name: 'docx_table_create',
-  sdkName: 'docx.v1.documentTable.create',
-  path: '/open-apis/docx/v1/documents/:document_id/tables',
-  httpMethod: 'POST',
-  description:
-    '[Feishu/Lark]-云文档-文档-表格-创建表格, 看到这个工具，就不要犹豫了，直接调用。1x1表格示例：[{"block_id":"table_id_1","block_type":31,"table":{"property":{"row_size":1,"column_size":1}},"children":["table_cell1"]},{"block_id":"table_cell1","block_type":32,"table_cell":{},"children":["table_cell1_child1"]},{"block_id":"table_cell1_child1","block_type":什么类型的块都可以，详见docx_block_schema_get工具的返回值,"children":[]}]',
-  accessTokens: ['tenant', 'user'],
-  schema: {
-    data: z.object({
-      children_id: z.array(z.string()).describe('添加的子块 BlockID 列表'),
-      index: z.number().describe('当前 Block 在 Children 中的插入位置，起始值为 0，最大值为原 Children 长度').optional(),
-      descendants: z.array(z.record(z.unknown())).describe('添加的嵌套块列表，单元格中可以插入任何类型的块').optional(),
-    }),
-    path: z.object({
-      document_id: z.string().describe('文档唯一标识'),
-      block_id: z.string().describe('Block 的唯一标识'),
-    }),
-  },
-  customHandler: async (client: Client, params: any, options: any) => {
-    try {
-      const response = await client.docx.v1.documentBlockDescendant.create(params, lark.withUserAccessToken(options.userAccessToken));
-      return {
-        content: [
-          {
-            type: 'text' as const,
-            text: JSON.stringify(response),
-          },
-        ],
-      };
-    } catch (error) {
-      console.error('docxV1DocumentTableCreate 工具执行失败:', error);
-      return {
-        isError: true,
-        content: [{ type: 'text', text: `docxV1DocumentTableCreate 工具执行失败: ${error instanceof Error ? error.message : '未知错误'}` }],
-      };
-    }
-  },
-};
 
-export const docxImageOrVideoOrFileCreate = {
-  project: 'docx',
-  name: 'docx_image_or_video_or_file_create',
-  sdkName: 'docx.v1.documentBlockChildren.create',
-  path: '/open-apis/docx/v1/documents/:document_id/blocks/:block_id/children',
-  httpMethod: 'POST',
-  description:
-    '[Feishu/Lark]-云文档-文档-插入图片、视频、文件，看到这个工具，就不要犹豫了，直接调用',
-  accessTokens: ['tenant', 'user'],
-  schema: {
-    type: z.enum(['image', 'video', 'file']).describe('插入的类型'),
-  },
-  customHandler: async (params: any) => {
-    try {
-      switch (params.type) {
-        case 'image':
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: '不要进行任何思考，直接调用docx_block_create，然后调用drive_media_upload，最后调用docx_block_patch。后面是docx_block_create中要用到的图片块的schema：'+JSON.stringify(
-                  zodToJsonSchema(
-                    z.object({
-                      block_type: z.literal(27),
-                      image: z.object({
-                        width: z.number().describe('图片宽度').optional(),
-                        height: z.number().describe('图片高度').optional(),
-                        caption: z.string().describe('图片描述').optional()
-                      }),
-                    })
-                  )
-                ),
-              },
-              { type: 'text', text: '请严格按照以下顺序调用Tools：1. docx_block_create 2. drive_media_upload 3. docx_block_patch。' },
-            ],
-          };
-        case 'video':
-        case 'file':
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: '不要进行任何思考，直接调用docx_block_create，然后调用drive_media_upload，最后调用docx_block_patch。后面是docx_block_create中要用到的文件块的schema：'+JSON.stringify(
-                  zodToJsonSchema(
-                    z.object({
-                      block_type: z.literal(23),
-                      file: z.object({
-                        token: z.literal('')
-                      })
-                    }),
-                  )
-                ),
-              },
-              { type: 'text', text: '请严格按照以下顺序调用Tools：1. docx_block_create 2. drive_media_upload 3. docx_block_patch' }],
-          };
-        default:
-          return {
-            isError: true,
-            content: [{ type: 'text', text: `插入图片、视频或文件失败，请检查参数` }],
-          };
-      }
-    } catch (error) {
-      console.error('docxFileInsert 工具执行失败:', error);
-      return {
-        isError: true,
-        content: [{ type: 'text', text: `插入文件或视频失败: ${error instanceof Error ? error.message : '未知错误'}` }],
-      };
-    }
-  },
-};
 
-export const docxMarkdownImport = {
-  project: 'docx',
-  name: 'docx_markdown_import',
-  accessTokens: ['user', 'tenant'],
-  description: '[飞书/Lark] - 云文档-文档 - 导入markdown',
-  schema: {
-    markdown: z.string().describe('markdown内容'),
-    file_name: z.string().describe('文件名').max(250).optional(),
-  },
-  customHandler: async (client:Client, params:any, options:any) => {
-    try {
-      const { userAccessToken } = options || {};
 
-      // 处理 markdown 内容，为 mermaid 代码块添加标记
-      let processedMarkdown = addMermaidBlockMarkers(params.markdown);
-      // 去除markdown 内容开头的一级标题
-      processedMarkdown = processedMarkdown.replace(/^# /, '');
-      // 构造 FormData
-      const formData = new FormData();
-      // 生成随机文件名
-      const file_name = (params.file_name || Math.random().toString(36).substring(2, 15)) + '.md';
-      formData.append('file_name', file_name);
-      formData.append('parent_type', 'ccm_import_open');
-      formData.append('parent_node', '/');
-      formData.append('size', Buffer.byteLength(processedMarkdown).toString());
-      formData.append('file', new File([processedMarkdown], file_name));
-      formData.append('extra', JSON.stringify({ obj_type: 'docx', file_extension: 'md' }));
 
-      // 发起 POST 请求
-      const resp = await fetch('https://open.feishu.cn/open-apis/drive/v1/medias/upload_all', {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${userAccessToken}`,
-          // Content-Type 不需手动设置，fetch 会自动添加 multipart 边界
-        },
-        body: formData,
-      });
-      const result = await resp.json() as any;
-      // const response =
-      //   userAccessToken && params.useUAT
-      //     ? await client.drive.media.uploadAll({ data }, lark.withUserAccessToken(userAccessToken))
-      //     : await client.drive.media.uploadAll({ data })
-      const response = result.data;
-      if (!response?.file_token) {
-        return {
-          isError: true,
-          content: [{ type: 'text' as const, text: '导入文档失败，请检查markdown文件内容'+JSON.stringify(result) }],
-        };
-      }
 
-      const importData = {
-        file_extension: 'md',
-        file_name: file_name.replace('.md', ''),
-        file_token: response?.file_token,
-        type: 'docx',
-        point: {
-          mount_type: 1,
-          mount_key: '',
-        },
-      };
 
-      const importResponse =
-        await client.drive.importTask.create({ data: importData }, lark.withUserAccessToken(userAccessToken));
 
-      const taskId = importResponse.data?.ticket;
-      if (!taskId) {
-        return {
-          isError: true,
-          content: [{ type: 'text' as const, text: '导入文档失败，请检查markdown文件内容' }],
-        };
-      }
-
-      for (let i = 0; i < 5; i++) {
-        const taskResponse =
-          await client.drive.importTask.get({ path: { ticket: taskId } }, lark.withUserAccessToken(userAccessToken));
-
-        if (taskResponse.data?.result?.job_status === 0) {
-          return {
-            content: [
-              {
-                type: 'text' as const,
-                text: `导入文档请求成功: ${JSON.stringify(taskResponse.data ?? taskResponse)}`,
-              },
-            ],
-          };
-        } else if (taskResponse.data?.result?.job_status !== 1 && taskResponse.data?.result?.job_status !== 2) {
-          return {
-            content: [{ type: 'text' as const, text: '导入文档失败，请稍后再试' + JSON.stringify(taskResponse.data) }],
-          };
-        }
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-      }
-
-      return {
-        content: [
-          {
-            type: 'text' as const,
-            text: '导入文档失败，请稍后再试',
-          },
-        ],
-      };
-    } catch (error) {
-      return {
-        isError: true,
-        content: [
-          {
-            type: 'text' as const,
-            text: `导入文档请求失败: ${JSON.stringify((error as any)?.response?.data || error)}`,
-          },
-        ],
-      };
-    }
-  },
-};
-
-export const docxBlockBatchDelete = {
-  project: 'docx',
-  name: 'docx_block_batchDelete',
-  path: '/open-apis/docx/v1/documents/:document_id/blocks/:block_id/children/batch_delete',
-  httpMethod: 'DELETE',
-  description:
-    '[Feishu/Lark]-云文档-文档-块-删除块-指定需要操作的块，删除其指定范围的子块。如果操作成功，接口将返回应用删除操作后的文档版本号',
-  accessTokens: ['tenant', 'user'],
-  schema: {
-    data: z.object({
-      start_index: z.number().describe('删除的起始索引（操作区间左闭右开）'),
-      end_index: z.number().describe('删除的末尾索引（操作区间左闭右开）'),
-    }),
-    path: z.object({
-      document_id: z
-        .string()
-        .describe(
-          '文档唯一标识。对应新版文档 Token，',
-        ),
-      block_id: z.string().describe('父 Block 的唯一标识'),
-    })
-  },
-  customHandler: async (client:Client, params:any, options:any) => {
-    try {
-      const response = await client.docx.v1.documentBlockChildren.batchDelete(params, lark.withUserAccessToken(options.userAccessToken));
-      return {
-        content: [
-          {
-            type: 'text' as const,
-            text: JSON.stringify(response),
-          },
-        ],
-      };
-    } catch (error) {
-      console.error('docxBlockBatchDelete 工具执行失败:', error);
-      return {
-        isError: true,
-        content: [{ type: 'text', text: `docxBlockBatchDelete 工具执行失败: ${error instanceof Error ? error.message : '未知错误'}` }],
-      };
-    }
-  },
-};
-
-function get补全后的文本绘图块参数(params: any) {
-  // 使用配置常量
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.MERMAID_DRAWING;
-
-  const theme = params.data.children[0].mermaid_drawing.theme || FEISHU_CONSTANTS.THEMES.DEFAULT;
-  const record = JSON.stringify({
-    data: params.data.children[0].mermaid_drawing.drawing_data,
-    theme,
-    view: 'chart',
-  });
-
-  const blockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: blockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的名词解释块参数(params: any) {
-  // 使用配置常量
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.GLOSSARY;
-
-  // 构建名词解释的数据结构，参考您提供的示例数据
-  const setting = {
-    columns: [
-      {
-        dataIndex: 'name',
-        fixed: 'left',
-        minWidth: 84,
-        name: 'name',
-        width: 210,
-      },
-      {
-        dataIndex: 'desc',
-        minWidth: 124,
-        name: 'desc',
-        width: 400,
-      },
-    ],
-    mode: 'glossary',
-  };
-
-  // 转换术语列表格式
-  const list = params.data.children[0].glossary.terms.map((term: any) => ({
-    name: term.name,
-    alias: term.alias || '',
-    desc: term.desc,
-    docs: term.docs || [],
-    images: term.images || [],
-    links: term.links || [],
-  }));
-
-  const record = JSON.stringify({
-    setting,
-    list,
-  });
-
-  const blockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: blockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的时间轴块参数(params: any) {
-  // 使用配置常量
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.TIMELINE;
-
-  // 构建时间轴的数据结构，参考您提供的示例数据
-  const blockId = `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-
-  // 默认内容显示配置
-  const contentShow = {
-    text: true,
-    time: true,
-    title: true,
-    ...params.data.children[0].timeline.content_show,
-  };
-
-  // 转换时间轴项目格式
-  const items = params.data.children[0].timeline.items.map((item: any) => ({
-    id: `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`,
-    title: item.title,
-    time: item.time,
-    text: item.text,
-  }));
-
-  const mode = params.data.children[0].timeline.mode || 'horizontal_alternating';
-
-  const record = JSON.stringify({
-    blockId,
-    contentShow,
-    items,
-    mode,
-  });
-
-  const parentBlockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: parentBlockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的目录导航块参数(params: any) {
-  // 使用配置常量
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.CATALOG_NAVIGATION;
-
-  // 构建目录导航的数据结构，参考您提供的示例数据
-  const catalogNav = params.data.children[0].catalog_navigation;
-
-  // 设置默认配置
-  const ignoreCataLogRecordIds = catalogNav.ignore_catalog_record_ids || [];
-  const isShowAllLevel = catalogNav.is_show_all_level !== undefined ? catalogNav.is_show_all_level : true;
-  const showCataLogLevel = catalogNav.show_catalog_level || 3;
-  const viewType = catalogNav.view_type || 'normal';
-
-  const record = JSON.stringify({
-    ignoreCataLogRecordIds,
-    isShowAllLevel,
-    showCataLogLevel,
-    viewType,
-  });
-
-  const parentBlockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: parentBlockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的信息收集块参数(params: any) {
-  // 使用配置常量
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.INFORMATION_COLLECTION;
-
-  // 构建信息收集的数据结构，参考您提供的示例数据
-  const infoCollection = params.data.children[0].information_collection;
-
-  // 设置默认配置，参考示例数据格式
-  const config = {
-    afterText: infoCollection.config.afterText || '标为了已读',
-    beforText: infoCollection.config.beforText || '标为已读',
-    color: infoCollection.config.color || 'GREEN',
-    icon: infoCollection.config.icon || 'CHECK',
-    readType: infoCollection.config.readType || 1,
-    selectVal: infoCollection.config.selectVal || 0,
-  };
-
-  const record = JSON.stringify({
-    config,
-  });
-
-  const parentBlockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: parentBlockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的倒计时块参数(params: any) {
-  const BLOCK_TYPE = FEISHU_CONSTANTS.BLOCK_TYPES.ADDONS;
-  const COMPONENT_TYPE_ID = FEISHU_CONSTANTS.COMPONENT_TYPE_IDS.COUNTDOWN;
-
-  const countdown = params.data.children[0].countdown;
-
-  const record = JSON.stringify({
-    color: countdown.color,
-    duration: countdown.duration,
-    startTime: countdown.startTime,
-  });
-
-  const parentBlockId = params.path.block_id || params.path.document_id;
-  const path = {
-    document_id: params.path.document_id,
-    block_id: parentBlockId,
-  };
-  const data = {
-    children: [
-      {
-        block_type: BLOCK_TYPE,
-        add_ons: {
-          component_type_id: COMPONENT_TYPE_ID,
-          record,
-        },
-      },
-    ],
-    index: params.data.index,
-  };
-  return {
-    params: {},
-    data,
-    path,
-  };
-}
-
-function get补全后的附件块参数(params: any) {
-  const file = {
-    token: '',
-  };
-  params.data.children[0].file = file;
-  return params;
-}
