+++ src/tools/document/media/index.ts
@@ -1,0 +1,1 @@
+export { docxImageOrVideoOrFileCreate } from './create';
