+++ src/tools/document/table/index.ts
@@ -1,0 +1,1 @@
+export { docxV1DocumentTableCreate } from './create';
